requirements.txt
spotipy==2.23.0
pandas==2.0.3
python-dotenv==1.0.1
requests==2.25.1
Flask==3.0.3
scikit-learn==0.24.1
